requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.2
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autopage==0.5.2
certifi==2025.8.3
cfgv==3.5.0
charset-normalizer==3.4.2
cliff==4.13.0
cmaes==0.12.0
cmd2==2.7.0
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distlib==0.4.0
dllist==2.0.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
hf-xet==1.1.7
huggingface-hub==0.34.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.6.15
idna==3.10
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.8
lightning==2.5.2
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
maxsim-cpu==0.1.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
optuna==2.10.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.5.0
prettytable==3.17.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.11.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.2
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.4
rich==14.2.0
rich-argparse==1.7.2
rootutils==1.0.7
safetensors==0.6.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.44
stack-data==0.6.3
stevedore==5.6.0
sympy==1.14.0
tensorrt==10.12.0.36
tensorrt_cu12==10.12.0.36
tensorrt_cu12_bindings==10.12.0.36
tensorrt_cu12_libs==10.12.0.36
timm==1.0.19
tomli==2.3.0
torch==2.8.0
torch_tensorrt==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.13
yarl==1.20.1
